chg: Admin renamed and refactored to IncidentManagementContainer
