--- app/containers/RepoListItem/Wrapper.js
@@ -1,10 +1,0 @@
-import styled from 'styled-components';
-
-const Wrapper = styled.div`
-  width: 100%;
-  height: 100%;
-  display: flex;
-  align-items: space-between;
-`;
-
-export default Wrapper;
