--- app/meldingen/incident-management/AdminComponent/messages.js
@@ -1,13 +1,0 @@
-/*
- * AdminComponent Messages
- *
- * This contains all the text for the AdminComponent component.
- */
-import { defineMessages } from 'react-intl';
-
-export default defineMessages({
-  header: {
-    id: 'app.components.AdminComponent.header',
-    defaultMessage: 'This is the AdminComponent component !',
-  },
-});
