--- app/meldingen/incident-management/AdminComponent/style.scss
@@ -1,3 +1,0 @@
-.admin-component {
-  min-height: 400px
-}
